solving conflict
