--- utils/helpers.ts
@@ -1,41 +1,0 @@
-require('dotenv').config();
-import { chromium, firefox} from 'playwright';
-
-export async function grabAllDomainCombos (obj: any){
-   let queue = [obj]
-   let values = []
-   while (queue.length > 0){
-     let current = queue.shift();
-     for (let val of Object.values(current)){
-       if(typeof val === "object") queue.push(val);
-       else values.push(val);
-     }
-   }
-   return values;
-}
-
-export async function markTestStatus( status, vPage ){
-  if (status = true) {
-    await vPage.evaluate(_ => {}, `browserstack_executor: ${JSON.stringify({action: 'setSessionStatus',arguments: {status: 'passed',reason: 'Test Passed'}})}`);
-  }
-  else if (status = null || false) {
-   await  vPage.evaluate(_ => {}, `browserstack_executor: ${JSON.stringify({action: 'setSessionStatus',arguments: {status: 'failed',reason: 'Test Failed'}})}`);
-
-  }
-}
-
-export async function instantiateBrowserstack( capabilities, vBrowser, testName ){
-  capabilities['client.playwrightVersion'] = '1.22.2';  // Playwright version being used on your local project needs to be passed in this capability for BrowserStack to be able to map request and responses correctly
-  console.log(process.env.BROWSERSTACK_USERNAME, process.env.BROWSERSTACK_ACCESS_KEY)
-  capabilities['browserstack.username'] = process.env.BROWSERSTACK_USERNAME;
-  capabilities['browserstack.accessKey'] = process.env.BROWSERSTACK_ACCESS_KEY;
-  capabilities['name'] = testName;
-  
-  console.log("Starting test -->", capabilities['name']);
-  console.log(`wss://cdp.browserstack.com/playwright?caps=${encodeURIComponent(JSON.stringify(capabilities))}`)
-  
-  vBrowser =  await chromium.connect({
-    wsEndpoint: `wss://cdp.browserstack.com/playwright?caps=${(encodeURIComponent(JSON.stringify(capabilities)))}`
-});
-  return vBrowser;
-}
